+++ src/types/JtConfig.ts
@@ -1,0 +1,7 @@
+export type JtExtraType = 'allow' | 'forbid' | 'ignore';
+
+export interface JtConfigType {
+  'extra'?: JtExtraType;
+  'frozen'?: boolean;
+  'strict'?: boolean;
+}